ArrayList and Autoboxing
Finished the coding challenges for ArrayList and Autoboxing lectures. Need to figure out how to get a little more throughput. It might be better to just go through all the lectures super quick and code as quickly as possible with low understanding so I get the "familiarity bump" with a large variety of subjects.